Update Batch-Process.ino
Changed StateMachine() to just repeat instructions when void loop() loops all the way around. Reasoning is that we might need to set variable to "correct" value if override in HMI has been set and then removed.

Counters and other instructions have been given a new variable called PreviouslyVisited, when a case is visited and loop sensitive instructions are present we set the variable true and reset it back to false whenever we break out of loop.