Finished working version before starting process of integrating dgraph.
Memory fragmentation issue we had in the past has been mitigated by removing
all use of global strings, replacing them with char arrays. Local strings
still used some places due to ArduinoJson library not liking to deserialize/
serialize char arrays. But local strings release all memory used after the
function they are a part of is complete, removing most problems with string
usage...

Due to memory usage, error handling has gone back to simple json
property instead of storage of last "good" value.. Deserialization error
message now displayed in the HMI layer (aka GUI).

Temporary solution for how to override updated sensor values. Must be
changed in the future to a better solution where one bind each override
button to a specific property, where we can see which sensor is under
override and must not be updated or at least ignored by Process Layer. For
Actuators this is not an issue as they are updated in the Control Layer,
where the information stream goes directly to the HMI Layer and we can
put overrides/simulate/change actuator states before they are sent to the
Process layer. Sensor values on the other hand are updated directly in the
Process layer (in theory, now it is of course hosted in Node-RED with HMI
layer) and therefore will reach the Control layer before having the chance
of being overrided/simulated/changed in the HMI layer. This is non ideal as
we sometime want to force a condition to happen by using sensor values and
therefore we must in the future probably bind each override check mark in GUI
with each corresponding sensor and then check which sensor is in override
mode and should not be updated in the Control layer. Maybe use binary instead
of json to save length of json character..?